+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Fix wrong typing for `R.sortByDescending` - [Issue #797](https://github.com/selfrefactor/rambda/issues/797)
 
+- Improve `R.mapParallelAsync` typings to allow optional `batchSize` parameter.
+
 10.3.3
 
 - Fix wrong typing for `R.reject` - [Issue #779](https://github.com/selfrefactor/rambda/issues/779)